# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: OpenAccounting API
  version: "1.0.0"
  description: |
    REST API for the OpenAccounting accounting platform.

    ## Authentication
    All `/api/v1/...` endpoints require a Bearer token:
    ```
    Authorization: Bearer oa_live_…
    ```

    ## Idempotency
    `POST` endpoints accept an `Idempotency-Key` header. Replays
    within 24 hours return the original response without re-executing
    the mutation. Keys are scoped per token + ledger.

    ## Pagination
    List endpoints return `{"data": [...]}` with a `Link: <…>; rel="next"`
    header when more results are available. Pass `?limit=` (default 50,
    max 200) and `?cursor=` from the previous `Link` header.

    ## Errors
    All errors return RFC 7807 problem-details JSON
    (`Content-Type: application/problem+json`).

    ## Incoming Events
    External systems can push signed events to
    `POST /api/events/{ledger_id}` using the ledger's
    `incoming_events_secret` (HMAC-SHA256 in the
    `X-OA-Event-Signature: sha256=<hex>` header).

    ## Automation Rules
    Ledger owners create rules via the web UI at
    `/ledgers/{id}/automations`. Rules map an allowlisted trigger
    (event type) + conditions → an action (webhook / email /
    categorise transaction) and execute on the jobs queue.
servers:
  - url: /
    description: Same origin
paths:
...
  /ledgers/{ledger_id}/invoices:
    get:
      operationId: listInvoices
      summary: List invoices in a ledger
      tags: [Invoices]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/PaginationParams"
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceList"
    post:
      operationId: createInvoice
      summary: Create an invoice (idempotent)
      tags: [Invoices]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /ledgers/{ledger_id}/invoices/{id}:
    get:
      operationId: getInvoice
      summary: Get a single invoice
      tags: [Invoices]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/InvoiceIdPath"
      responses:
        "200":
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          $ref: "#/components/responses/NotFound"
  /ledgers/{ledger_id}/invoices/{id}/mark-paid:
    post:
      operationId: markInvoicePaid
      summary: Mark an invoice as paid (idempotent)
      tags: [Invoices]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/InvoiceIdPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Invoice marked as paid
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /ledgers/{ledger_id}/invoices/{id}/void:
    post:
      operationId: voidInvoice
      summary: Void an invoice with a reason (idempotent)
      tags: [Invoices]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/InvoiceIdPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Invoice voided
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
...
  /ledgers/{ledger_id}/contacts:
    get:
      operationId: listContacts
      summary: List contacts in a ledger
      tags: [Contacts]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/PaginationParams"
      responses:
        "200":
          description: Contact list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactList"
    post:
      operationId: createContact
      summary: Create a contact (idempotent)
      tags: [Contacts]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /ledgers/{ledger_id}/contacts/{id}:
    get:
      operationId: getContact
      summary: Get a single contact
      tags: [Contacts]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/ContactIdPath"
      responses:
        "200":
          description: Contact details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateContact
      summary: Update a contact
      tags: [Contacts]
      parameters:
        - $ref: "#/components/parameters/LedgerIdPath"
        - $ref: "#/components/parameters/ContactIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactBody"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    LedgerIdPath:
      name: ledger_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    InvoiceIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ContactIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Client-generated key for idempotent POSTs (24 h window)
    PaginationParams:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 200
          default: 50
      - name: cursor
        in: query
        required: false
        schema:
          type: string
  schemas:
...
    Invoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ledger_id:
          type: string
          format: uuid
        status:
          type: string
        total:
          type: number
        amount_paid:
          type: number
        invoice_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
    InvoiceList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
    CreateInvoiceBody:
      type: object
      required: [invoice_date, due_date, total, lines]
      properties:
        invoice_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        total:
          type: number
        description:
          type: string
        lines:
          type: array
          items:
            type: object
...
    Contact:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ledger_id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        kind:
          type: string
    ContactList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
    CreateContactBody:
      type: object
      required: [name, kind]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        kind:
          type: string
          enum: [customer, vendor, both]
    UpdateContactBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        kind:
          type: string
          enum: [customer, vendor, both]
...